Modularize db
