+++ Server/server.js
@@ -1,6 +1,7 @@
 require("dotenv").config();
 const express = require("express");
 const mysql = require("mysql");
+const https = require("node:https");
 
 // AUTH0 REQUIRED IMPORTS
 const session = require("express-session");
@@ -29,68 +30,6 @@ app.use(
 
 // MAIN LOGIN LOGIC -- NEEDS TO BE TESTED
 app.use(routes);
-
-//
-const connection = mysql.createConnection({
-  host: "localhost",
-  port: "3306",
-  user: "dev",
-  password: "password",
-  database: "albertslist",
-});
-
-// Test API call
-app.get("/api", (req, res) => {
-  res.json({ users: ["user1", "user2", "user3"] });
-});
-
-// Test DB Call
-app.get("/api/users", (req, res) => {
-  connection.connect();
-
-  connection.query("SELECT * from Users", (err, result) => {
-    if (err) throw err;
-
-    res.json(result);
-  });
-  connection.end();
-});
-
-// Test Insert User
-app.post("/api/post/insertuser", (req, res) => {
-  var username = req.body.username;
-  var password = req.body.password;
-  var firstname = req.body.firstname;
-  var lastname = req.body.lastname;
-  var email = req.body.email;
-  var isPartner = req.body.isPartner;
-  var zip = req.body.zip;
-
-  connection.connect();
-  var sql =
-    "INSERT INTO Users (username, password, firstname,lastname,email,isPartner,zip) VALUES('" +
-    username +
-    "','" +
-    password +
-    "','" +
-    firstname +
-    "','" +
-    lastname +
-    "','" +
-    email +
-    "','" +
-    isPartner +
-    "','" +
-    zip +
-    "')";
-  connection.query(sql, (err, result) => {
-    if (err) throw err;
-    res.send({
-      message: "New user:" + username + " was added to the database",
-    });
-  });
-  connection.end();
-});
 
 app.listen(PORT, (error) => {
   if (!error)
